Move transitionCase into useCase hook
