+++ apps/judicial-system/web/src/utils/hooks/useCase.tsx
@@ -1,16 +1,19 @@
 import { useMutation } from '@apollo/client'
 import {
   Case,
+  CaseState,
+  CaseTransition,
   NotificationType,
   SendNotificationResponse,
   UpdateCase,
 } from '@island.is/judicial-system/types'
 import {
   SendNotificationMutation,
+  TransitionCaseMutation,
   UpdateCaseMutation,
 } from '@island.is/judicial-system-web/graphql'
 import { CreateCaseMutation, CreateCourtCaseMutation } from '../mutations'
-import { parseString } from '../formatters'
+import { parseString, parseTransition } from '../formatters'
 
 type autofillProperties = Pick<
   Case,
@@ -38,6 +41,11 @@ const useCase = () => {
   const [createCaseMutation, { loading: isCreatingCase }] = useMutation(
     CreateCaseMutation,
   )
+  const [
+    transitionCaseMutation,
+    { loading: isTransitioningCase },
+  ] = useMutation(TransitionCaseMutation)
+
   const [
     createCourtCaseMutation,
     { loading: creatingCourtCase },
@@ -120,6 +128,49 @@ const useCase = () => {
     }
   }
 
+  const transitionCase = async (
+    workingCase: Case,
+    setWorkingCase: React.Dispatch<React.SetStateAction<Case | undefined>>,
+  ) => {
+    if (!workingCase) {
+      return false
+    }
+
+    switch (workingCase.state) {
+      case CaseState.NEW:
+        try {
+          // Parse the transition request
+          const transitionRequest = parseTransition(
+            workingCase.modified,
+            CaseTransition.OPEN,
+          )
+
+          const { data } = await transitionCaseMutation({
+            variables: { input: { id: workingCase.id, ...transitionRequest } },
+          })
+
+          if (!data) {
+            return false
+          }
+
+          setWorkingCase({
+            ...workingCase,
+            state: data.transitionCase.state,
+          })
+
+          return true
+        } catch (e) {
+          return false
+        }
+      case CaseState.DRAFT:
+      case CaseState.SUBMITTED:
+      case CaseState.RECEIVED:
+        return true
+      default:
+        return false
+    }
+  }
+
   const updateCase = async (id: string, updateCase: UpdateCase) => {
     // Only update if id has been set
     if (!id) {
@@ -180,6 +231,8 @@ const useCase = () => {
     autofill,
     createCourtCase,
     creatingCourtCase,
+    transitionCase,
+    isTransitioningCase,
   }
 }
